requirements.txt
flask>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
authlib>=1.3.0